New Store Release Link
Adds a new section for the Edge link and also adjusts the Chrome store link to the redirected location now that is it published (saving the world 1 less request at a time). In the same spirit I added a trailing slash to the root domain.
Fixes a case typo in options.js.